refactor: Relocate queries to proper folder
